third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/unicode-max-length-for-device-name-namePrefix.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
promise_test(() => {
  // \u2764's UTF-8 respresentation is 3 bytes long.
  // 81 chars * 3 bytes/char = 243 bytes
  let unicode_name = generate_string(81, '\u2764');
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{namePrefix: unicode_name}]}),
    new DOMException(
      'Failed to execute \'requestDevice\' on \'Bluetooth\': ' +
      'A \'name\' or \'namePrefix\' longer than 240 bytes ' +
      'results in no devices being found, because a device can\'t ' +
      'acquire a name longer than 240 bytes.',
    'NotFoundError'));
}, 'Unicode string with utf8 representation between (240, 248] bytes in ' +
   '\'namePrefix\' must throw NotFoundError.');
</script>
